requirements.txt
# Data & APIs
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0

# ML
scikit-learn>=1.3.0
tensorflow>=2.13.0
joblib>=1.3.0

# Feature importance
shap>=0.43.0

# Storage
pymongo>=4.5.0

# Web app
streamlit>=1.28.0

# Config
python-dotenv>=1.0.0

# Plotting (EDA & dashboard)
matplotlib>=3.7.0
plotly>=5.18.0

# API (per PDF: Streamlit + Flask/FastAPI)
flask>=3.0.0
